requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
grpcio==1.74.0
grpcio-tools==1.74.0
pandas==2.3.1
